pyproject.toml
[project]
name = "smart-ffmpeg"
version = "0.1.0"
description = "A CLI tool to generate and execute FFmpeg commands from natural language."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "rich",
    "openai",
    "python-dotenv",
]

[project.scripts]
smart-ffmpeg = "smart_ffmpeg:main"
